rbd: do not execute rbd sparsify when volume is in use
This is a backport PR.
This commit makes sure sparsify() is not run when rbd
image is in use.
Running rbd sparsify with workload doing io and too
frequently is not desirable.
When a image is in use fstrim is run and sparsify will
be run only when image is not mapped.

Signed-off-by: Yati Padia <ypadia@redhat.com>